Move ConnectionAuthorizationType to network/auth
This location makes more sense to the use of
ConnectionAuthorizationType
