+++ libsplinter/src/service/network/interconnect/mod.rs
@@ -429,9 +429,9 @@ pub mod tests {
     use std::collections::VecDeque;
 
     use crate::mesh::{Envelope, Mesh};
+    use crate::network::auth::ConnectionAuthorizationType;
     use crate::network::connection_manager::{
-        AuthorizationResult, Authorizer, AuthorizerError, ConnectionAuthorizationType,
-        ConnectionManager,
+        AuthorizationResult, Authorizer, AuthorizerError, ConnectionManager,
     };
     use crate::network::dispatch::{
         dispatch_channel, ConnectionId, DispatchError, DispatchLoopBuilder, Dispatcher, Handler,
